dir and user are now separated
